Table: move TableNG utils into a utils/ folder
utils.ts had grown to 1773 lines covering a dozen unrelated concerns. Split it
into six modules under utils/ with a barrel index.ts, so every existing
`./utils` / `../utils` import site resolves unchanged:

  fields   field and cell-option accessors, display names, alignment
  rows     frame-to-rows conversion, comparators, sorting, filtering, row keys
  pills    pill inference and the bounded cache backing it
  height   typography contexts, text measurers, row and footer heights
  width    fixed and content-aware column widths
  display  cell colors, data links, display values, inspect formatting

The modules form a DAG: fields and pills have no intra-folder dependencies; rows,
display and height build on those; width sits on top.

Modules inside the folder import their siblings by direct path rather than
through the barrel — routing them through index.ts would pull the whole folder
into every leaf and recreate the Cells/renderers cycle internally.

Also moves ApplyFilterResult from utils.ts into types.ts, which removes a real
types.ts <-> utils.ts cycle, since types.ts imported the interface while utils.ts
imported the rest of types.ts.

This commit deliberately leaves utils.test.ts untouched: the existing 2989-line
suite passing against the new module graph is the evidence that the move changes
no behaviour. The tests are split in the follow-up commit.
