+++ packages/grafana-ui/src/components/Table/TableNG/utils/index.ts
@@ -1,0 +1,9 @@
+// Barrel for the TableNG table utilities. Re-exports only — grafana-ui ships `sideEffects: false`.
+// Modules in this folder import their siblings by direct path rather than through here, so the
+// barrel never pulls the whole folder (and the Cells/renderers cycle) into a single leaf.
+export * from './display';
+export * from './fields';
+export * from './height';
+export * from './pills';
+export * from './rows';
+export * from './width';
